refactor: core module test classes
