+++ core/build.gradle
@@ -37,12 +37,5 @@ android {
 dependencies {
     api project(":base")
     api Dependencies.androidxCoreKtx
-
     implementation Dependencies.coroutinesAndroid
-
-    testImplementation(Dependencies.mockK)
-    testImplementation(Dependencies.coroutinesTest)
-
-    testImplementation(platform(Dependencies.junitBom))
-    testImplementation(Dependencies.junitJupiter)
 }
